+++ src/components/box/clipboard-history.tsx
@@ -1,46 +1,54 @@
 import { t } from "@/i18n";
 import { createDialog } from "../dialogs/dialog";
-import { createSignal, For } from "solid-js";
+import {
+  Accessor,
+  createEffect,
+  createSignal,
+  For,
+} from "solid-js";
 import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
 } from "../ui/tooltip";
 import { toast } from "solid-sonner";
-
+import { SendClipboardMessage } from "@/libs/core/messge";
+import { createTimeAgo } from "@/libs/utils/timeago";
 export const createClipboardHistoryDialog = () => {
-  const {
-    open: openDialog,
-    close,
-    submit,
-    Component,
-  } = createDialog({
+  const { open: openDialog, Component } = createDialog({
     title: () => t("common.clipboard_history.title"),
     content: () => (
-      <div class="max-h-[60vh] overflow-y-auto">
-        <ul class="flex flex-col gap-2">
+      <div class="max-h-[60vh] overflow-y-auto p-2">
+        <ul class="flex flex-col-reverse gap-2">
           <For each={clipboardCacheData()}>
             {(item) => (
               <Tooltip>
                 <TooltipTrigger
                   as="li"
-                  class="line-clamp-2 cursor-pointer overflow-hidden
-                    whitespace-pre-wrap text-wrap rounded-md border
-                    border-border p-1 text-sm hover:bg-muted"
-                >
-                  <li
-                    onClick={() => {
-                      navigator.clipboard &&
-                        navigator.clipboard.writeText(item).then(() => {
-                          toast.success(t("common.notification.copy_success"));
+                  class="flex cursor-pointer flex-col rounded-md border border-border
+                    p-1 text-sm hover:bg-muted"
+                  onClick={() => {
+                    navigator.clipboard &&
+                      navigator.clipboard
+                        .writeText(item.data)
+                        .then(() => {
+                          toast.success(
+                            t(
+                              "common.notification.copy_success",
+                            ),
+                          );
                         });
-                    }}
-                  >
-                    {item}
-                  </li>
+                  }}
+                >
+                  <p class="line-clamp-2 overflow-hidden whitespace-pre-wrap text-wrap">
+                    {item.data}
+                  </p>
+                  <p class="self-end text-xs text-muted-foreground">
+                    {createTimeAgo(item.createdAt)}
+                  </p>
                 </TooltipTrigger>
                 <TooltipContent class="whitespace-pre-wrap">
-                  {item}
+                  {item.data}
                 </TooltipContent>
               </Tooltip>
             )}
@@ -51,10 +59,14 @@ export const createClipboardHistoryDialog = () => {
   });
 
   const [clipboardCacheData, setClipboardCacheData] =
-    createSignal<string[]>([]);
+    createSignal<SendClipboardMessage[]>([]);
 
-  const open = (clipboardCacheData: string[]) => {
-    setClipboardCacheData(clipboardCacheData);
+  const open = (
+    clipboardCacheData: Accessor<SendClipboardMessage[]>,
+  ) => {
+    createEffect(() => {
+      setClipboardCacheData(clipboardCacheData());
+    });
     openDialog();
   };